Make zone & vspan updates more generic
