+++ services/ui/src/app/disruption/components/disruption.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { Category, TimeRegionSchema, TimePointSchema, MultiVariateTimeSeriesData, Annotations, BaseAnnotation, VSpan, Zone, TimePoint, TimeRegion} from "@/types"
+import { Category, TimeRegionSchema, TimePointSchema, MultiVariateTimeSeriesData, Annotations, VSpan, Zone, ZoneSchema, VSpanSchema, DisplayAnnotation} from "@/types"
 import { ZoneProvider } from "@/app/components/providers/zone-provider"
 import { VSpanProvider } from "@/app/components/providers/vpsan-provider"
 import { DisruptionTable } from "./disruption-table"
@@ -8,6 +8,29 @@ import { ContextMenuProvider } from "@/app/components/providers/context-menu-pro
 import { TimeSeries } from "@/app/components/plots/time-series"
 import { Zones } from "@/app/components/tools/zones"
 import { VSpans } from "@/app/components/tools/vspans"
+import { createAnnotationToDisplayAnnotationFunc, updateAnnotations } from "@/app/utils"
+
+const disruptionCategories: Category[] = [
+    { label: "Disruption", color: 'rgb(255, 0, 0)' },
+]
+
+const zoneCategories: Category[] = [
+    { label: "RampUp", color: 'rgb(233, 170, 98)' },
+    { label: "FlatTop", color: 'rgb(120, 167, 85)' },
+    { label: "RampDown", color: 'rgb(108, 189, 224)' }
+]
+
+const zoneCategoryColors = zoneCategories.reduce<Record<string, string>>((acc, curr) => {
+  acc[curr.name] = curr.color;
+  return acc;
+}, {});
+
+const disruptionCategoryColors = disruptionCategories.reduce<Record<string, string>>((acc, curr) => {
+  acc[curr.name] = curr.color;
+  return acc;
+}, {});
+
+const colorMapping = { ...disruptionCategoryColors, ...zoneCategoryColors };
 
 type DisruptionViewInfo = {
   data: MultiVariateTimeSeriesData,
@@ -15,28 +38,21 @@ type DisruptionViewInfo = {
   setAnnotations: (annotations: Annotations) => void
 };
 
+
 export const DisruptionView = ({data, annotations, setAnnotations} : DisruptionViewInfo) => {
-    const disruptionCategories: Category[] = [
-            { label: "Disruption", color: 'rgb(255, 0, 0)' },
-        ]
+    const convertAnnotationToDisplayAnnotation = createAnnotationToDisplayAnnotationFunc(colorMapping);
+    const displayAnnotations: DisplayAnnotation[] = annotations.map(convertAnnotationToDisplayAnnotation);
+    const zones: Zone[] = displayAnnotations.filter((x: DisplayAnnotation) => ZoneSchema.safeParse(x).success);
+    const vspans: VSpan[] = displayAnnotations.filter((x: DisplayAnnotation) => VSpanSchema.safeParse(x).success);
 
-    const convertTimePointToVSpan = (item: TimePoint) => {
-        const category = disruptionCategories.find((x: Category) => x.label === item.label);
-        return {x: item.time, category: category};
-    };
-    const spans = annotations.filter((x: BaseAnnotation) => TimePointSchema.safeParse(x).success).map(convertTimePointToVSpan);
+    const updateZones = (newZones: Array<Zone>) => {
+        updateAnnotations(setAnnotations, newZones, TimeRegionSchema);
+    }
 
-    const zoneCategories: Category[] = [
-            { label: "RampUp", color: 'rgb(233, 170, 98)' },
-            { label: "FlatTop", color: 'rgb(120, 167, 85)' },
-            { label: "RampDown", color: 'rgb(108, 189, 224)' }
-        ]
+    const updateVSpans = (newVSpans: Array<VSpan>) => {
+        updateAnnotations(setAnnotations, newVSpans, TimePointSchema);
+    }
 
-    const convertTimeRegionToZone = (item: TimeRegion) => {
-        const category = zoneCategories.find((x: Category) => x.label === item.label);
-        return {x0: item.time_min, x1: item.time_max, category: category};
-    };
-    const zones = annotations.filter((x: BaseAnnotation) => TimeRegionSchema.safeParse(x).success).map(convertTimeRegionToZone);
 
     const plotData: Plotly.Data[] = [
         {
@@ -87,39 +103,11 @@ export const DisruptionView = ({data, annotations, setAnnotations} : DisruptionV
         dragmode: 'pan',
     };
 
-    const updateZones = (newZones: Array<Zone>) => {
-        let zones = newZones.map(item => (
-            TimeRegionSchema.parse({
-                time_min: item.x0,
-                time_max: item.x1,
-                label: item.category.label
-            })
-        ));
-
-
-        const otherAnnotations: BaseAnnotation = annotations.filter((x: BaseAnnotation) => !TimeRegionSchema.safeParse(x).success);
-        zones = zones.concat(otherAnnotations);
-        setAnnotations(zones);
-    };
-
-    const updateVSpans = (newVSpans: Array<VSpan>) =>  {
-        let spans = newVSpans.map(item => (
-            TimePointSchema.parse({
-                time: item.x,
-                label: item.category.label
-            })
-        ));
-
-
-        const otherAnnotations: Annotations = annotations.filter((x: BaseAnnotation) => !TimePointSchema.safeParse(x).success);
-        spans = spans.concat(otherAnnotations);
-        setAnnotations(spans);
-    };
     
     return (
         <div className="flex flex-col items-center space-y-3">
             <ContextMenuProvider menuId="disruption-menu">
-                <VSpanProvider categories={disruptionCategories} initialData={spans} onModifyVSpan={updateVSpans}>
+                <VSpanProvider categories={disruptionCategories} initialData={vspans} onModifyVSpan={updateVSpans}>
                     <ZoneProvider categories={zoneCategories} initialData={zones} onModifyZone={updateZones}>
                         <TimeSeries plotId="Disruption" plotConfig={{data: plotData, layout: plotLayout}}>
                             <Zones />
