+++ services/ui/src/app/elms/components/elms.tsx
@@ -1,18 +1,37 @@
 'use client'
-import { BaseAnnotation, MultiVariateTimeSeriesData, TimeRegion, Zone, Category } from "@/types"
+import { MultiVariateTimeSeriesData, Zone, Category, DisplayAnnotation, ZoneSchema, TimeRegionSchema, Annotation } from "@/types"
 import { ZoneProvider } from "@/app/components/providers/zone-provider"
 import { ContextMenuProvider } from "@/app/components/providers/context-menu-provider"
 import { TimeSeries } from "@/app/components/plots/time-series"
 import { Zones } from "@/app/components/tools/zones"
 import 'react-contexify/ReactContexify.css';
 import Plotly from "plotly.js-dist";
+import { createAnnotationToDisplayAnnotationFunc, updateAnnotations } from "@/app/utils"
+
+const zoneCategories: Category[] = [
+    { name: "ELM", color: 'rgb(233, 170, 98)' },
+    { name: "H-Mode", color: 'rgb(100, 170, 98)' },
+]
+
+const zoneCategoryColors = zoneCategories.reduce<Record<string, string>>((acc, curr) => {
+  acc[curr.name] = curr.color;
+  return acc;
+}, {});
 
 type ELMViewInfo = {
     data: MultiVariateTimeSeriesData, 
-    annotations: BaseAnnotation, 
-    setAnnotations: (annotations: BaseAnnotation) => void
+    annotations: Annotation, 
+    setAnnotations: (annotations: Annotation) => void
 };
+
 export const ELMView = ({data, annotations, setAnnotations}: ELMViewInfo) => {
+    const convertAnnotationToDisplayAnnotation = createAnnotationToDisplayAnnotationFunc(zoneCategoryColors);
+    const displayAnnotations: DisplayAnnotation[] = annotations.map(convertAnnotationToDisplayAnnotation);
+    const zones: Zone[] = displayAnnotations.filter((x: DisplayAnnotation) => ZoneSchema.safeParse(x).success);
+
+    const updateZones = (newZones: Array<Zone>) => {
+        updateAnnotations(setAnnotations, newZones, TimeRegionSchema);
+    }
 
     var dataTrace = {
         name: 'Dalpha',
@@ -173,26 +192,6 @@ export const ELMView = ({data, annotations, setAnnotations}: ELMViewInfo) => {
     };
 
     
-    const zoneCategories: Category[] = [
-        { name: "ELM", color: 'rgb(233, 170, 98)' },
-        { name: "H-Mode", color: 'rgb(100, 170, 98)' },
-    ]
-
-    const convertRegionToZone = (item: TimeRegion) => {
-        const category = zoneCategories.find(x => x.name === item.label);
-        return {x0: item.time_min, x1: item.time_max, category: category};
-    };
-    const zones = annotations.map(convertRegionToZone);
-
-    const updateAnnotations = (newZones: Array<Zone>) => {
-        const zones = newZones.map(item => ({
-                time_min: item.x0,
-                time_max: item.x1,
-                label: item.category.name
-        }));
-
-        setAnnotations(zones);
-    }
 
     return (
         <div className="flex space-y-3">
